+++ pages/create.tsx
@@ -1,5 +1,11 @@
-import { SdNewImagePrompt } from '../components/SdNewImagePrompt';
+import { Title } from "@mantine/core";
+import { SdNewImagePrompt } from "../components/SdNewImagePrompt";
 
 export default function CreatePrompt() {
-  return <SdNewImagePrompt />;
+  return (
+    <div className="container">
+      <Title order={1}>test a prompt</Title>
+      <SdNewImagePrompt />
+    </div>
+  );
 }